Moved ::types::errors module to error::vec_elem
