+++ backend/libs/errors/src/lib.rs
@@ -5,6 +5,7 @@ mod initialize;
 mod object;
 mod parse;
 mod status;
+mod vec_elem;
 mod websocket;
 
 pub use attempt::MaximumAttemptExceeded;
@@ -14,4 +15,5 @@ pub use initialize::InitError;
 pub use object::ObjectNotFound;
 pub use parse::ParseError;
 pub use status::StatusFailure;
+pub use vec_elem::{RawVecElemErrs, VecElementErr, VecElementErrs};
 pub use websocket::WebsocketError;
