+++ backend/libs/exchanges/binance/notification/src/entities/account_update.rs
@@ -1,8 +1,8 @@
-use std::num::ParseFloatError;
+use ::std::num::ParseFloatError;
 
-use mongodb::bson::DateTime;
-use serde::{Deserialize, Serialize};
-use types::errors::{RawVecElemErrs, VecElementErr, VecElementErrs};
+use ::errors::{RawVecElemErrs, VecElementErr, VecElementErrs};
+use ::mongodb::bson::DateTime;
+use ::serde::{Deserialize, Serialize};
 
 use ::types::casting::cast_datetime_from_i64;
